Text encoding fix
<rant>
How happy UTF-8 can be!
Through Athens ISO 8859-1 is thought as fair as it.
But what of that? The characters are ruined
It will destroy all characters that it doesn't know
And as it errors, making databases incorrect
So UTF-8, having the correct characters,
Things base and vile, having double encoded
Functions can transform from UTF-8 and ISO 8859-1
PHP looks not with the encoding but with the characters
And therefore is winged Cupid painted blind             <-- couldn't think of a parody
Nor hath hath Love's mind of any judgment taste         <-- again

ARRGH THE REST OF THAT IS SO LONG I'M QUITTING
</rant>