Simple templates (FINALLY!)
Here, you can transclude only (substitution may be coming soon) by
typing {{pagename}}. The software currently swallows up all the text
after a vertical bar. You can work around like this:

 * Create Template:! with '|'
 * You can call with {{vertical{{!}}bar}} now

Here's how it works internally:

Text is parsed:
 {{foo{{bar}}}}
 ^^   ^^   ^^^^
 |    |    | |
 |    |    | +- Two closing braces, depth = 0 after change, recording stopped.
 |    |    + Two closing braces, depth = 1, recording not stopped yet
 |    +- Two braces, depth = 2
 +- Two braces, depth = 0, start recording, depth = 1

That way we can get something like:
array [
  {
    text: "foo{{bar}}",
    start: 0,
    end: 14
  }
]

Then, we invoke the actual parser. A callback is supplied which
returns the page name or NULL if the page doesn't exist.

So, now we do:
{{foo{{bar}}}}
|            |
|            +- Position 14
+- Position 0

The expression foo{{bar}} is evaluated, then the callback
is supplied the remaining. Then, substr_replace is called
to replace the template, and the remaining offsets after
it are adjusted.

If the resulting string still has templates (parseTemplates()
returns an array with items), it is parsed over and over again.

But you could just do {{pagename}} without reading all of this
and still get it.