<noinclude> tags and a few magic words
You can use a few magic words now. They
should be understood by the parser. The magic words are:

 - {{BASEPAGENAME}}: {{FULLPAGENAME}} minus the namespace
 - {{FULLPAGENAME}}: the full page name

Noinclude tags provide content that shouldn't be
included with the template. For example:

<noinclude>foo</noinclude>bar

will display 'foobar' on the template page, but
will display 'bar' on the pages it is used on.

You can also pass arguments:
{{template|foo=bar}}

Then, {{argfoo}} will be 'bar' when transcluded. However,
on the template page, {{argfoo}} is undefined.